Addressing ReverseTabnabbing with info links
Adding noopener, it ensures that the linked page does not have access to window.opener from the source page. While noreferrer makes sure that the request referer header is not sent along with the request. This way, the destination site does not see the originating URL that the user is coming from. The destination can not then manipulate original page to show phishing content.

https://www.comparitech.com/blog/information-security/reverse-tabnabbing/